factor out parts of the fuzzers into separate files
